pyproject.toml
[tool.poetry]
name = "uav-object-counting-bsc-cs"
version = "0.1.0"
description = ""
authors = ["Aleksa Sibinović <as1871@student.uni-lj.si>"]
packages = [
  { include = "uav_object_counting_bsc_cs" }
]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "*"
torch = "*"
torchvision = "*"
ultralytics = "*"
numpy = "*"
pyyaml = "*"
hydra-core = "*"
matplotlib = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
